+++ lib/common.ts
@@ -1,10 +1,6 @@
 export type AnyRecord = Record<string, any>;
 
-export type Meta = {
-  key: string;
-  name: string;
-  description: string;
-};
+
 
 export type BaseRequestConfig = {
   url: string;
@@ -15,28 +11,41 @@ export type GetRequestConfig = BaseRequestConfig & {
   method: "GET";
 };
 
+export type MultipartBodyConfig = {
+    key:string
+    fileName:string;
+    fileValue:ArrayBuffer;
+    contentType:string
+}
+
 export type MutatingRequestConfig = BaseRequestConfig & {
   method: "POST" | "PATCH" | "PUT" | "DELETE";
   jsonBody?: any;
-  multipartBody?: any;
+  multipartBody?: MultipartBodyConfig[];
 };
 
 export type RequestConfig = GetRequestConfig | MutatingRequestConfig;
 
-export type ExecuteServiceError = {
-  stringError(message: string): never;
-  cause(message: string): never;
-};
-
-export type RequestResult<Response = any> = {
+export type RequestResult<Response = ArrayBuffer> = {
   status: number;
-  response: Readonly<Partial<Response>>;
+  response: Readonly<Response> | undefined;
 };
-
-export type ExecuteService = {
+export type ExecuteServiceAuth = {
+  hook: (
+    fn: (url: string, headers: Record<string, string>) => string,
+    guid: string,
+    timeout: number
+  ) => string | undefined;
+};
+export type ExecuteServiceUtils = {
   base64Encode: (input: string) => string;
   base64Decode: (input: string) => string;
-  request: <Response>(config: RequestConfig) => RequestResult<Response>;
-  error: ExecuteServiceError;
+  stringError(message: string): never;
   console: (input: string) => void;
 };
+
+export type ExecuteServiceNetwork = {
+  request: <Response>(config: RequestConfig) => RequestResult<Response>;
+};
+
+export type ExecuteService = ExecuteServiceAuth & ExecuteServiceNetwork & ExecuteServiceUtils;