+++ lib/index.ts
@@ -3,9 +3,9 @@ export * from "./debugging";
 export * from "./common";
 export * from "./connection";
 export * from "./integration";
-
+export * from "./utils/api/HttpClient";
 import type { Integration } from "./integration";
 
 declare global {
-  var integration: Integration;
+  var app: Integration;
 }